+++ routes/testimonials.php
@@ -19,6 +19,7 @@ switch ($_SERVER['REQUEST_METHOD']) {
     case 'DELETE':
         $matches = [];
         if (preg_match('/^\/api\/testimonials\/(\d+)$/', $_SERVER['REQUEST_URI'], $matches)) {
+             requireSuperAdmin();
             $testimonialController->delete($matches[1]);
         } else {
             http_response_code(400);
